requirements.txt
Django==1.6.2
pyquery==1.2.8
requests==2.2.1
factory-boy==2.3.1
